% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_test.R
\name{generate_selector}
\alias{generate_selector}
\title{Generate CSS Selector from Element Description}
\usage{
generate_selector(element, type = "input")
}
\arguments{
\item{element}{Character string describing the element}

\item{type}{Character string indicating element type}
}
\value{
Character string containing CSS selector
}
\description{
Generate CSS Selector from Element Description
}
\keyword{internal}
