% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_test.R
\name{generate_test_code}
\alias{generate_test_code}
\title{Convert Test Actions to Shinytest2 Code}
\usage{
generate_test_code(actions)
}
\arguments{
\item{actions}{List of parsed test actions}
}
\value{
Character string containing generated test code
}
\description{
Convert Test Actions to Shinytest2 Code
}
\keyword{internal}
